+++ src/files/index.spec.js
@@ -1,5 +1,6 @@
 import Files from './index';
 import * as objectUtils from '../utils/objects';
+import * as paginationUtils from '../utils/pagination';
 
 describe('Files', function() {
   let baseRequest;
@@ -49,6 +50,51 @@ describe('Files', function() {
 
     it('appends the supplied sdk to the class instance', function() {
       expect(files._sdk).to.deep.equal(baseSdk);
+    });
+  });
+
+  describe('delete', function() {
+    context('the file ID is provided', function() {
+      let expectedFile;
+      let promise;
+      let request;
+
+      beforeEach(function() {
+        expectedFile = fixture.build('file');
+
+        request = {
+          ...baseRequest,
+          delete: this.sandbox.stub().resolves()
+        };
+
+        const files = new Files(baseSdk, request);
+        files._baseUrl = expectedHost;
+
+        promise = files.delete(expectedFile.id);
+      });
+
+      it('deletes the file from the server', function() {
+        return promise.then(() => {
+          expect(request.delete).to.be.calledWith(
+            `${expectedHost}/files/${expectedFile.id}`
+          );
+        });
+      });
+
+      it('returns a fulfilled promise', function() {
+        return expect(promise).to.be.fulfilled;
+      });
+    });
+
+    context('the file ID is not provided', function() {
+      it('returns a rejected promise with an error', function() {
+        const files = new Files(baseSdk, baseRequest);
+        const promise = files.delete();
+
+        return expect(promise).to.be.rejectedWith(
+          'A file ID is required to delete a file'
+        );
+      });
     });
   });
 
@@ -181,6 +227,97 @@ describe('Files', function() {
           'A file ID is required for getting information about a file'
         );
       });
+    });
+  });
+
+  describe('getAll', function() {
+    let filesFromServerAfterFormat;
+    let filesFromServerBeforeFormat;
+    let formatPaginatedDataFromServer;
+    let filesFiltersAfterFormat;
+    let filesFiltersBeforeFormat;
+    let promise;
+    let request;
+    let toSnakeCase;
+
+    beforeEach(function() {
+      filesFromServerAfterFormat = {
+        _metadata: fixture.build('paginationMetadata'),
+        records: fixture.buildList(
+          'file',
+          faker.random.number({ min: 1, max: 10 })
+        )
+      };
+      filesFromServerBeforeFormat = {
+        ...filesFromServerAfterFormat,
+        records: filesFromServerAfterFormat.records.map((asset) =>
+          fixture.build('file', asset, { fromServer: true })
+        )
+      };
+      filesFiltersBeforeFormat = {
+        limit: faker.random.number({ min: 10, max: 1000 }),
+        offset: faker.random.number({ max: 1000 }),
+        orderBy: faker.random.arrayElement([
+          'content_type',
+          'created_at',
+          'description',
+          'filename',
+          'id',
+          'organization_id',
+          'owner_id',
+          'status',
+          'updated_at'
+        ]),
+        reverseOrder: faker.random.boolean(),
+        status: faker.random.arrayElement(['ACTIVE', 'UPLOADING'])
+      };
+      filesFiltersAfterFormat = {
+        ...filesFiltersBeforeFormat
+      };
+
+      formatPaginatedDataFromServer = this.sandbox
+        .stub(paginationUtils, 'formatPaginatedDataFromServer')
+        .returns(filesFromServerAfterFormat);
+      request = {
+        ...baseRequest,
+        get: this.sandbox.stub().resolves(filesFromServerBeforeFormat)
+      };
+      toSnakeCase = this.sandbox
+        .stub(objectUtils, 'toSnakeCase')
+        .returns(filesFiltersAfterFormat);
+
+      const files = new Files(baseSdk, request);
+      files._baseUrl = expectedHost;
+
+      promise = files.getAll(filesFiltersBeforeFormat);
+    });
+
+    it('formats the filter options', function() {
+      return promise.then(() => {
+        expect(toSnakeCase).to.be.calledWith(filesFiltersBeforeFormat);
+      });
+    });
+
+    it('gets a list of files from the server', function() {
+      return promise.then(() => {
+        expect(request.get).to.be.calledWith(`${expectedHost}/files`, {
+          params: filesFiltersAfterFormat
+        });
+      });
+    });
+
+    it('formats the files data', function() {
+      return promise.then(() => {
+        expect(formatPaginatedDataFromServer).to.be.calledWith(
+          filesFromServerBeforeFormat
+        );
+      });
+    });
+
+    it('returns a list of files', function() {
+      return expect(promise).to.be.fulfilled.and.to.eventually.equal(
+        filesFromServerAfterFormat
+      );
     });
   });
 });